+++ integration/spark/app/src/test/java/io/openlineage/spark/agent/ColumnLineageIntegrationTest.java
@@ -52,7 +52,7 @@ public class ColumnLineageIntegrationTest {
   private static Network network = Network.newNetwork();
   private static SparkSession spark;
   public static final int POSTGRES_PORT = 5432;
-  private static final int MOCKSERVER_PORT = 1082;
+  private static final int MOCKSERVER_PORT = 1083;
 
   private static int mappedPort;
   private static ClientAndServer mockServer;